refactor: useHighlight훅 분리
하이라이트에 따른 blockList 상태, 버튼 이벤트 핸들러들을 useHighlight훅으로 분리
