+++ frontend/src/components/highlight/HighlightEditor.tsx
@@ -1,13 +1,5 @@
-import { useState } from 'react';
-
-import {
-  getRemovedHighlightList,
-  getSelectionInfo,
-  getSelectionOffsetInBlock,
-  getUpdatedBlockByHighlight,
-  removeSelection,
-  splitTextWithHighlightList,
-} from '@/utils';
+import { useHighlight } from '@/hooks';
+import { splitTextWithHighlightList } from '@/utils';
 
 interface Highlight {
   start: number;
@@ -24,12 +16,7 @@ interface HighlightEditorProps {
 }
 
 const HighlightEditor = ({ text }: HighlightEditorProps) => {
-  const [blockList, setBlockList] = useState<Block[]>(() =>
-    text.split('\n').map((text) => ({
-      text,
-      highlightList: [],
-    })),
-  );
+  const { blockList, handleClickHighlight, handleClickHighlightRemover } = useHighlight({ text });
 
   /**
    * 블록의 값에 따라 하이라이트 여부를 반영한 태그들을 렌더링하는 함수
@@ -53,68 +40,6 @@ const HighlightEditor = ({ text }: HighlightEditorProps) => {
         ))}
       </>
     );
-  };
-
-  const handleClickHighlight = () => {
-    const info = getSelectionInfo();
-    if (!info) return;
-    const { selection, startBlock, startBlockIndex, endBlock, endBlockIndex } = info;
-
-    const newBlockList = blockList.map((block, index) => {
-      if (index < startBlockIndex) return block;
-      if (index > endBlockIndex) return block;
-      if (index === startBlockIndex) {
-        const startOffset = getSelectionOffsetInBlock({ selection, blockElement: startBlock });
-        return getUpdatedBlockByHighlight({ blockIndex: index, ...startOffset, blockList });
-      }
-      if (index === endBlockIndex) {
-        const endOffset = getSelectionOffsetInBlock({ selection, blockElement: endBlock });
-        return getUpdatedBlockByHighlight({ blockIndex: index, start: 0, end: endOffset.start, blockList });
-      }
-      return {
-        ...block,
-        highlightList: [{ start: 0, length: block.text.length }],
-      };
-    });
-
-    setBlockList(newBlockList);
-    removeSelection();
-  };
-
-  const handleClickHighlightRemover = () => {
-    const info = getSelectionInfo();
-    if (!info) return;
-    const { selection, startBlock, startBlockIndex, endBlock, endBlockIndex } = info;
-
-    const newBlockList = blockList.map((block, index) => {
-      if (index < startBlockIndex) return block;
-      if (index > endBlockIndex) return block;
-      if (index === startBlockIndex) {
-        const startOffset = getSelectionOffsetInBlock({ selection, blockElement: startBlock });
-        return {
-          ...block,
-          highlightList: getRemovedHighlightList({ highlightList: block.highlightList, ...startOffset }),
-        };
-      }
-      if (index === endBlockIndex) {
-        const endOffset = getSelectionOffsetInBlock({ selection, blockElement: endBlock });
-        return {
-          ...block,
-          highlightList: getRemovedHighlightList({
-            highlightList: block.highlightList,
-            start: 0,
-            end: endOffset.start,
-          }),
-        };
-      }
-      return {
-        ...block,
-        highlightList: [],
-      };
-    });
-
-    setBlockList(newBlockList);
-    removeSelection();
   };
 
   return (
