+++ frontend/src/hooks/highlight/index.tsx
@@ -1,0 +1,1 @@
+export { default as useHighlight } from './useHighlight';
